added_tags

+++ api/v1/controller/tags.py
@@ -6,12 +6,13 @@ from sqlalchemy import *
 from sqlalchemy.orm import *
 from flask import abort, jsonify, make_response, request
 from api.v1.controller.path import app_controller
+from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
 from models import *
 
 def serialized_tags(tag):
     """Serialize a tag object to a JSON-serializable format."""
     return {
-        'id': tag.id,
+        'tag_id': tag.tagid,
         'title': tag.title,
         'slug': tag.slug,
         'created': tag.created,
@@ -27,7 +28,7 @@ def get_posts_by_tagId(tag_id):
     of a specific tags, or a specific user
     """
     list_tags = []
-    tag = Tag.query.filter_by(id=tag_id).first()
+    tag = Tag.query.filter_by(tagid=tag_id).first()
     if not tag:
         abort(404)
     for post in tag.posts:
@@ -43,37 +44,53 @@ def get_tags():
     serialized_tag = [serialized_tags(tag) for tag in tags]
     return jsonify(serialized_tag)
 
-@app_controller.route('/tags/<int:id>', methods=['GET'], strict_slashes=False)
+@app_controller.route('/tags/<id>', methods=['GET'], strict_slashes=False)
 def get_tag_by_id(id):
     """Create a new view for tags object that handles
     all default RESTFul API actions:"""
     # users = []
-    tag = Tag.query.filter_by(id=id).first()
+    tag = Tag.query.filter_by(tagid=id).first()
     if tag is None:
         abort(404)
     serialized_tag = serialized_tags(tag)
     return jsonify(serialized_tag)
 
-@app_controller.route('/tags/<int:id>', methods=['DELETE'],
+@app_controller.route('/tags/<id>', methods=['DELETE'],
                  strict_slashes=False)
+@jwt_required()
 def delete_tag(id):
     """Create a new view for tag object that
     handles all default RESTFul API actions:"""
-    tag = Tag.query.filter_by(id=id).first()
+    current_user = get_jwt_identity()
+    if not current_user:
+        return jsonify({"message": "Not a valid user"}), 401
+    user = User.query.filter_by(email=current_user).first()
+    if not user:
+        return jsonify({"message": "Not a valid user"}), 401
+    
+    tag = Tag.query.filter_by(tagid=id).first()
     if tag is None:
         abort(404)
     db.session.delete(tag)
     db.session.commit()
     return (jsonify({}))
 
-@app_controller.route('posts/<int:post_id>/tags/<int:tag_id>', methods=['POST'], strict_slashes=False)
+@app_controller.route('posts/<post_id>/tags/<tag_id>', methods=['POST'], strict_slashes=False)
+@jwt_required()
 def create_posts_tags(post_id, tag_id):
     """Create a new view for tag object
     that handles all default RESTFul API actions:"""
-    post = Post.query.filter_by(id=post_id).first()
+    current_user = get_jwt_identity()
+    if not current_user:
+        return jsonify({"message": "Not a valid user"}), 401
+    user = User.query.filter_by(email=current_user).first()
+    if not user:
+        return jsonify({"message": "Not a valid user"}), 401
+
+    post = Post.query.filter_by(postid=post_id).first()
     if not post:
         abort(404)
-    tag = Tag.query.filter_by(id=tag_id).first()
+    tag = Tag.query.filter_by(tagid=tag_id).first()
     if not tag:
         abort(404)
     if tag in post.tags:
@@ -87,9 +104,17 @@ def create_posts_tags(post_id, tag_id):
 
 
 @app_controller.route('/tags', methods=['POST'], strict_slashes=False)
+@jwt_required()
 def post_tag():
     """Create a new view for Tag object
     that handles all default RESTFul API actions:"""
+    current_user = get_jwt_identity()
+    if not current_user:
+        return jsonify({"message": "Not a valid user"}), 401
+    user = User.query.filter_by(email=current_user).first()
+    if not user:
+        return jsonify({"message": "Not a valid user"}), 401
+
     if not request.get_json():
         return make_response(jsonify({'error': 'Not a JSON'}), 400)
     if 'title' not in request.get_json():
@@ -99,12 +124,20 @@ def post_tag():
     db.session.commit()
     return make_response(jsonify(serialized_tags(tag)), 201)
 
-@app_controller.route('/tags/<int:id>', methods=['PUT'],
+@app_controller.route('/tags/<id>', methods=['PUT'],
                  strict_slashes=False)
+@jwt_required()
 def put_tag(id):
     """Create a new view for tag object that
     handles all default RESTFul API actions:"""
-    tag = Tag.query.filter_by(id=id).first()
+    current_user = get_jwt_identity()
+    if not current_user:
+        return jsonify({"message": "Not a valid user"}), 401
+    user = User.query.filter_by(email=current_user).first()
+    if not user:
+        return jsonify({"message": "Not a valid user"}), 401
+
+    tag = Tag.query.filter_by(tagid=id).first()
     if tag is None:
         abort(404)
     if not request.get_json():
